JPMS compatible schema output paths.
JPMS requires resources that are accessible from outside the module to be in _unique_ and _open_ packages.

To achieve this, we introduce a `OutputLocationStrategy` option to the generator, which defaults to placing the generated schema in a directory tree that matches the package layout of the source types.

This means, if the `--output-directory` is set to the resource root of a project, the schema files are located in the same package as the source types in the resulting jar file.